refactor: disable unused features to get minimal version (#2)
* docs: added partial README, and LICENSE to the project

* chore: release v0.1.2-beta.0

* refactor: removed next-auth and any implementation

This commit removed next-auth to the project as we will not be needing it. This includes removing it
on tRPC. This also removes its dependencies.

* refactor: disabled prisma from tRPC routes

This commit disabled, removed any prisma related codes in the project, specifically the tRPC routes
and routers. This removed the codes, primarily commenting it out, but leaves the dependencies
attached for future reimplementation.

* refactor: removed boilerplate homepage and added basic home component

* chore: release v0.1.2-beta.1

---------

Co-authored-by: GitHub Actions <41898282+github-actions[bot]@users.noreply.github.com>